Massive restructure
I was not satisfied with the structure of the old code.  I reorganized
it, documented it, and reworked RandomPlayer per Peter Taylor's example.
There is now a package, com.sadakatsu.clue.ai, that contains an abstract
SpeedClueAI class upon which other AIs can be (relatively easily) built.
RandomPlayer is now RandomAI, a subclass of SpeedClueAI, and the
resulting JAR file is now randomAI.jar.  In addition, the Ant build
script generates com.sadakatsu.clue that can be used to build one's own
AIs and Javadocs that can be found in the doc directory.