Detailed rules, minor test server changes
I have rewritten the contest rules to have an easier introduction for
new players.  "overview.md" will be posted as an edit to the
StackExchange question, which has links to more markdown files in this
repo for further details.

In addition, I have made some changes to the TestServer.  It now formats
its output more cleanly so it is easier to read how a match proceeds.
In addition, it automatically routes each launched program's output to a
file named using the program identifier.