`ContestServer` and other core code changes
The biggest change is the addition of the `ContestServer` class.  It is
what will actually run the tournament.  `contestServer.jar` is run using
the syntax `java -jar contestServer.jar entries gameCount`, where
"entries" holds all the AI launch scripts and "gameCount" is the number
of games each permutation of AIs will play against each other.  For the
official tournament, I will run this as `java -jar contestServer.jar
entries.txt 100`.

To make `ContestServer` work, and removing code duplications, there has
been a lot of internal restructuring in the code.  This involves changes
to the way `TestServer` works and its output, but these changes should
not be too confusing.

As before, all of the code I have written in the core is available in
`com.sadakatsu.clue.jar`.